requirements.txt
streamlit == 0.82.0
yfinance == 0.1.59
plotly == 4.14.3
ta == 0.7.0
pandas == 1.2.1
numpy == 1.19.5
mpmath == 1.1.0
python-dateutil == 2.8.1
Pillow == 8.0.1
google-cloud-firestore == 2.1.1
json5 == 0.9.5
google-auth == 1.30.0
